Website Update
Created an event listener that runs an anonymous function that creates a variable that stores an object collection of all select elements nested within the govLinks form. Also created a for loop that loops through the allSelect object collection, and for each selection list in the collection, creates an anonymous function that changes the page shown in the browser window to the value of the target of the event object that initiated the onchange event.